refactor: 파일구조 변경, component>shared 폴더 제작
