+++ src/components/shared/SearchBar/SearchBar.js
